Break out the `package ...` command hierarchy
Continuing with the clean-up of `pkg/cmd/pulumi`, this factors out the
`package ...` hierarchy.
