+++ pkg/cmd/pulumi/packagecmd/package_pack_sdk.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package packagecmd
 
 import (
 	"context"
